Simplify Camera Add nested_for_each for nested for loops Move Rectify from Graphics to global function in Rect.h

+++ Engine/Font.h
@@ -11,8 +11,8 @@ public:
 	Font( const TextFormat::Properties &Props );
 
 	Sizei GetCharSize()const;
-	size_t MaxCharsPerRow( size_t LineWidth )const;
-	size_t MaxCharsPerColumn( size_t ColumnHeight )const;
+	int MaxCharsPerRow( int LineWidth )const;
+	int MaxCharsPerColumn( int ColumnHeight )const;
 
 	void DrawChar( float X, float Y, char C, Color Clr, Graphics &Gfx )const;
 	void DrawChar( const Vec2f &Pos, char C, Color Clr, Graphics &Gfx )const;
